requirements.txt
# Core
openai>=1.0.0             # For Grok API / OpenRouter
python-dotenv>=1.0.0       # Load .env variables

# RAG / embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn>=1.2.2
numpy>=1.25.0

# PDF & document processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Backend / frontend
streamlit>=1.25.0
requests>=2.31.0

# Optional: for BM25 / text re-ranking
rank_bm25>=0.2.2
